src/cloudfunctions/fetch_ergast_files/requirements.txt
requests==2.30.0
google-cloud-storage==2.9.0